refactor: add advanceSettingScreenHelper.js file
